TvojeMama1: lekce GIT, Added Mirek, MozillaUkol moved
